+++ developing/frontend/kitoAi/src/app/features/home/services/home.service.ts
@@ -26,8 +26,7 @@ export class HomeService {
 					} else {
 						return data as Array<Home>;
 					}
-				}),
-				catchError(this.utils.handleError)
+				})
 			);
 	}