+++ developing/frontend/kitoAi/src/app/features/layout/base.component.ts
@@ -1,0 +1,17 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'kito-base',
+  standalone: true,
+  imports: [CommonModule],
+  template: `
+    <p>
+      base works!
+    </p>
+  `,
+  styles: ``
+})
+export class BaseComponent {
+
+}